requirements.txt
boto3
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
